+++ CHANGELOG.md
@@ -10,6 +10,7 @@ title: Changelog
 
 * i18n: simplify formatter system and add JSDoc
 * i18n: remove country flags (it makes no sense for a language)
+* `<cc-button>`: Add a circle form when in hide-text with an image
 
 ## 6.4.0 (2021-04-21)